requirements.txt
python-docx>=0.8.11
# Alternative to textract for DOC files
pypdf2>=2.0.0
docx2txt>=0.8
python-pptx>=0.6.21
google-generativeai>=0.3.0
python-dotenv>=1.0.0
